[FIRRTL] Migrate GCT-SM to LowerAnnotations
Migrate annotation parsing/scattering related to Grand Central Signal
Mappings (GCT-SM) from the parser (FIRAnnotations.cpp) to the
LowerAnnotations pass.

This necessarily makes a number of superficial modifications to the
existing annotation encoding.  Specifically, any recursively scattered
annotations are changed to use "suffix" class names (".module" and
".target") that tag the module that participates in a GCT-SM module and
the source/sink targets of this GCT-SM.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>